feat(Smith): Moves the smith functions.
